+++ lib/event_bus/event_bus_plus.dart
@@ -1,0 +1,3 @@
+// library event_bus_plus;
+//
+export 'res/res.dart';