+++ lib/event_bus/res/res.dart
@@ -1,0 +1,4 @@
+export 'app_event.dart';
+export 'event_bus.dart';
+export 'history_entry.dart';
+export 'subscription.dart';