+++ lib/helpers/helper.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 import 'dart:developer';
-import 'package:event_bus_plus/event_bus_plus.dart';
+import 'package:nylo_support/event_bus/event_bus_plus.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';